chore: Add Poetry as new build manager
Poetry is now used as the build manager for the project. It controls
dependencies and installs (kept controlled with a .lock file) and builds
a venv automatically stored in .venv directory. Pyenv is necessary for
installing specific versions of python, which then uses .python-version
locally to run the right version.