fix(formatting): Correct string capitalisation
The description and body fields were being modifed to drop their
capitals throughout the text when capitalising the first letter. This
was behaviour of the built in str.capitalize() method. A new utility
function has been created which performs the desired behaviour of purely
upper casing the first character in a string.

Fix #25
Fix #28